+++ src/control_plane/tls/mod.rs
@@ -1,0 +1,10 @@
+#[cfg(test)]
+mod tests;
+
+mod generate_local_certificates;
+mod local_certificate_bundle;
+mod local_certificate_error;
+
+pub(crate) use generate_local_certificates::generate_local_certificates;
+pub(crate) use local_certificate_bundle::LocalCertificateBundle;
+pub(crate) use local_certificate_error::LocalCertificateError;